fix(stm32): discover Arduino_Core_STM32 framework libraries (SPI, Wire, ...)
The STM32 orchestrator installs the STM32duino core but never walked
libraries/*/ to surface bundled libraries to sketches. Any sketch that
transitively included <SPI.h> (e.g. FastLED's STM32 fastspi path) failed
with "fatal error: SPI.h: No such file or directory" despite SPI living
right there in the cache.

Mirror PR #164's Teensy fix:

- Add a shared FrameworkLibrary model + discover_framework_libraries()
  walker in fbuild-packages, used by both TeensyCores and Stm32Cores.
- Extract the #include-scanning / transitive-closure resolver out of the
  Teensy orchestrator into crate::framework_libs so both platforms share
  one code path.
- Wire Stm32Cores::get_framework_libraries() +
  get_framework_library_include_dirs() into the STM32 orchestrator after
  SrcWrapper so the resolver picks up SPI/Wire/EEPROM/... on demand and
  their include dirs reach the sketch compile command.

Fixes #202
