+++ crates/fbuild-packages/src/library/mod.rs
@@ -12,6 +12,7 @@ pub mod cmsis_framework;
 pub mod esp32_framework;
 pub mod esp32_platform;
 pub mod esp8266_framework;
+pub mod framework_library;
 pub mod library_compiler;
 pub mod library_downloader;
 pub mod library_info;
@@ -38,6 +39,7 @@ pub use cmsis_framework::CmsisFramework;
 pub use esp32_framework::Esp32Framework;
 pub use esp32_platform::Esp32Platform;
 pub use esp8266_framework::Esp8266Framework;
+pub use framework_library::FrameworkLibrary;
 pub use library_manager::LibraryResult;
 pub use library_spec::LibrarySpec;
 pub use nrf52_core::Nrf52Cores;
@@ -47,4 +49,4 @@ pub use sam_core::SamCores;
 pub use samd_core::SamdCores;
 pub use silabs_core::SilabsCores;
 pub use stm32_core::Stm32Cores;
-pub use teensy_core::{TeensyCores, TeensyFrameworkLibrary};
+pub use teensy_core::TeensyCores;
